fix(pr): Fix broken tests and conditional for deserializer
