+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/jackson/SecurityGroupRuleDeserializer.kt
@@ -8,14 +8,5 @@ import com.netflix.spinnaker.keel.jackson.PropertyNamePolymorphicDeserializer
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.jackson.JsonComponent
 
-@JsonComponent
-@ConditionalOnMissingBean(SecurityGroupRuleDeserializer::class)
-class SecurityGroupRuleDeserializer :
-  PropertyNamePolymorphicDeserializer<SecurityGroupRule>(SecurityGroupRule::class.java) {
-  override fun identifySubType(fieldNames: Collection<String>): Class<out SecurityGroupRule> =
-    when {
-      "blockRange" in fieldNames -> CidrRule::class.java
-      "account" in fieldNames -> CrossAccountReferenceRule::class.java
-      else -> ReferenceRule::class.java
-    }
-}
+abstract class SecurityGroupRuleDeserializer :
+  PropertyNamePolymorphicDeserializer<SecurityGroupRule>(SecurityGroupRule::class.java)
